requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
mysql-connector-python==9.4.0
pydantic==1.10.11
